feat: reusable Ask and Edit controllers based off Anthropic
